+++ raktarmozgas/MozgasTipus.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace raktarmozgas
+{
+    enum MozgasTipus //A készletmozgás lehetséges típusai
+    {
+        BESZERZES,
+        ELADAS,
+        VISSZARU,
+        SELEJT
+    }
+}
