+++ raktarmozgas/Program.cs
@@ -8,319 +8,6 @@ namespace raktarmozgas
 {
     internal class Program
     {
-        static readonly int termekFajta = 10; //Termékféleségek aktuális darabszáma
-        static readonly int mozgasFajta = 4; //Készletmozgás lehetséges típusainak száma
-
-        enum MozgasTipus //A készletmozgás lehetséges típusai
-        {
-            BESZERZES,
-            ELADAS,
-            VISSZARU,
-            SELEJT
-        }
-        struct KeszletMozgas //Egy készletmozgást reprezentáló struktúra
-        {
-            byte id;
-            byte ora;
-            byte perc;
-            string termek;
-            float mennyiseg;
-            double egysegAr;
-            MozgasTipus tipus;
-            string partner;
-
-            public float Mennyiseg {  get => mennyiseg; }
-            public double EgysegAr { get => egysegAr; }
-            public MozgasTipus Tipus { get => tipus; }
-            public string Partner { get => partner; }
-
-            public static List<KeszletMozgas> mozgasok = new(); //A készletmozgásokat reprezentáló struktúrák listája
-
-            /*Feldarabolja a beolvasott sorokat a megadott elválasztó jel mentén, az értékeket a struktúra változóiba tölti.*/
-            public static KeszletMozgas CreateKeszletMozgas(string input)
-            {
-                string[] data = input.Split(";");
-                string[] ido = data[1].Split(":");
-
-                KeszletMozgas keszletMozgas = new()
-                {
-                    id = byte.Parse(data[0]),
-                    ora = byte.Parse(ido[0]),
-                    perc = byte.Parse(ido[1]),
-                    termek = data[2],
-                    mennyiseg = float.Parse(data[3]),
-                    egysegAr = double.Parse(data[4]),
-                    tipus = (MozgasTipus)Enum.Parse(typeof(MozgasTipus), data[5]),
-                    partner = data[6]
-                };
-
-                return keszletMozgas;
-            }
-
-            /*Létrehoz egy tömböt a különböző féle termékek számára.*/
-            public static string[] GetProducts(List<KeszletMozgas> mozgas)
-            {
-                int i = 0;
-                string[] termekek = new string[termekFajta];
-
-                foreach (var item in mozgas)
-                {
-                    if (!termekek.Contains(item.termek))
-                    {
-                        termekek[i] = item.termek;
-                        i++;
-                    }
-                }
-                return termekek;
-            }
-
-            /*Létrehoz egy kétdimenziós tömböt a beszállított és eladott termékek mennyiségei számára (2. dimenzió a mozgástípus értéke).*/
-            public static float[,] GetProductAmounts(List<KeszletMozgas> mozgas)
-            {
-                string[] termekek = GetProducts(mozgas);
-                float[,] mennyisegek = new float[termekFajta, mozgasFajta];
-
-                foreach (var item
-                    in mozgas)
-                {
-                    int j = 0;
-                    while (j < termekek.Length && !(item.termek == termekek[j])) j++;
-                    mennyisegek[j, (int)item.tipus] += item.mennyiseg;
-                }
-                return mennyisegek;
-            }
-
-            /*Létrehoz egy kétdimenziós tömböt a termékek beszerzési és eladási ára számára (2. dimenzió a mozgástípus értéke).*/
-            public static double[,] GetProductPrices(List<KeszletMozgas> mozgas)
-            {
-                string[] termekek = GetProducts(mozgas);
-                double[,] termekArak = new double[termekFajta, mozgasFajta];
-
-                foreach (var item in mozgas)
-                {
-                    int j = 0;
-                    while (j < termekek.Length && !(item.termek == termekek[j])) j++;
-                    termekArak[j, (int)item.tipus] = item.egysegAr;
-                }
-                return termekArak;
-            }
-
-            /*Létrehoz egy tömböt az egyes termékek beszállítói számára (2. dimenzió a mozgástípus értéke)*/
-            public static string[,] GetProductShippers(List<KeszletMozgas> mozgas)
-            {
-                string[] termekek = GetProducts(mozgas);
-                string[,] beszallitok = new string[termekFajta, mozgasFajta];
-
-                foreach (var item in mozgas)
-                {
-                    int j = 0;
-                    while (j < termekek.Length && !(item.termek == termekek[j])) j++;
-                    beszallitok[j, (int)item.tipus] = item.Partner;
-                }
-                return beszallitok;
-            }
-
-            /*Óránként összegzi a ki- és bemenő forgalmakat.*/
-            public static int[,] SalesPerHour(List<KeszletMozgas> mozgas)
-            {
-                int[,] forgalmak = new int[24, mozgasFajta];
-
-                foreach (var item in mozgas)
-                {
-                    forgalmak[item.ora, (int)item.tipus]++;
-                }
-                return forgalmak;
-            }
-
-            /*Kiszámolja mely termékek készlete esett 50% alá, és visszaadja azon termékeket, melyekből rendelést kell leadni.*/
-            public static string[] OutOfStock(List<KeszletMozgas> mozgas)
-            {
-                string[] termekek = GetProducts(mozgas);
-                float[,] mennyisegek = GetProductAmounts(mozgas);
-                string[] kifogyoTermekek = new string[termekFajta];
-                int j = 0;
-
-                for (int i = 0; i < termekek.Length; i++)
-                {
-                    if (mennyisegek[i, (int)MozgasTipus.ELADAS] > mennyisegek[i, (int)MozgasTipus.BESZERZES] / 2)
-                    {
-                        kifogyoTermekek[j] = termekek[i];
-                        j++;
-                    }
-                }
-                return kifogyoTermekek;
-            }
-        }
-
-        struct Partner //Egy ügyfelet reprezentáló struktúra
-        {
-            int id = 0;
-            string nev = "";
-            MozgasTipus tipus = 0;
-            float mennyiseg = 0;
-            double ertek = 0;
-
-            public string Nev { get => nev; }
-            public float Mennyiseg { get => mennyiseg; }
-            public double Ertek { get => ertek; }
-
-            public static List<Partner> partnerek = new(); //Az ügyfeleket reprezentáló struktúrák listája
-
-            static Partner maxertek;
-            public static Partner maxErtek
-            {
-                get => maxertek;
-                set => maxertek = value;
-            }
-            public static Partner maxmenny;
-            public static Partner maxMenny
-            {
-                get => maxmenny;
-                set => maxmenny = value;
-            }
-            public Partner() { }
-
-            /*A konstruktor metódus overloadjában létrehoz két Partner példányt, melyek az adatok betöltése után visszaadják a legnagyobb
-             beszállítók adatait. (Az ügyfelek listájába nem töltődnek be!)*/
-            Partner(List<Partner> partnerek, string arg = "mennyiseg")
-            {
-                Partner max = (arg == "ertek") ? partnerek.MaxBy(e => e.ertek) : partnerek.MaxBy(m => m.mennyiseg);
-                id = max.id;
-                nev = max.nev;
-                MozgasTipus tipus = max.tipus;
-                mennyiseg = max.mennyiseg;
-                ertek = max.ertek;
-            }
-
-            /*Létrehozza az üzleti partnereket reprezentáló struktúrát.*/
-            static Partner CreatePartner(List<KeszletMozgas> mozgas, int index, int id) 
-            {
-                Partner partner = new()
-                {
-                    id = id+1,
-                    nev = mozgas[index].Partner,
-                    tipus = mozgas[index].Tipus,
-                    mennyiseg = mozgas[index].Mennyiseg,
-                    ertek = mozgas[index].EgysegAr * mozgas[index].Mennyiseg
-                };
-                return partner; 
-            }
-
-            /*Ha ugyanazon partnerhez kapcsolódó eseményt talál, összegzi a partnerhez kötődő árumogás mennyiségét és értékét, majd
-             frissíti a partnert reprezentáló struktúrát az aktuális értékekkel.*/
-            static Partner UpdatePartner(List<KeszletMozgas> mozgas, int index, int id)
-            {
-                Partner partner = new()
-                {
-                    id = id,
-                    nev = mozgas[index].Partner,
-                    tipus = mozgas[index].Tipus,
-                    mennyiseg = partnerek[id].mennyiseg + mozgas[index].Mennyiseg,
-                    ertek = partnerek[id].ertek + mozgas[index].EgysegAr * mozgas[index].Mennyiseg
-                };
-                return partner;
-            }
-
-            /*Partnerek szerint csoportosítja a készletmozgást, összegzi az egy üzleti partnerre vetített áruk mennyiségét és értékét.
-             Ezután a konstruktor overload metódusának meghívásával kiválasztja a legnagyobb mennyiségben és értékben szállító partnereket.*/
-            public static List<Partner> CreatePartnerList(List<KeszletMozgas> mozgas)
-            {
-               partnerek.Add(CreatePartner(mozgas, 0, 0));
-
-                for (int i = 1; i < mozgas.Count; i++)
-                {
-                    int j = 0;
-                    while (j < partnerek.Count && partnerek[j].nev != mozgas[i].Partner) j++;
-                    if (j == partnerek.Count) { partnerek.Add(CreatePartner(mozgas, i, j)); }
-                    else { partnerek.Add(UpdatePartner(mozgas, i, j)); partnerek.RemoveAt(j); }
-                }
-
-                maxertek = new Partner(partnerek, "ertek");
-                maxmenny = new Partner(partnerek, "mennyiseg");
-
-                return partnerek;
-            }
-
-            /*Kiszámolja az össz napi termékforgalom mennyiségét.*/
-            public static float[] SumTradeFlow(List<Partner> partnerek)
-            {
-                float[] osszmenny = new float[mozgasFajta];
-
-                foreach (var item in partnerek)
-                {
-                    osszmenny[(int)item.tipus] += item.mennyiseg;
-                }
-                return osszmenny;
-            }
-            /*Kiszámoljaaz össz napi pénzforgalom értékét.*/
-            public static double[] SumCashFlow(List<Partner> partnerek)
-            {
-                double[] osszertek = new double[mozgasFajta];
-
-                foreach (var item in partnerek)
-                {
-                    osszertek[(int)item.tipus] += item.ertek;
-                }
-                return osszertek;
-            }
-        }   
-
-        struct Termek //Egy termék állapotát reprezentáló struktúra
-        { 
-            int id; 
-            byte ora;
-            byte perc;
-            MozgasTipus tipus;
-            string nev;
-            float mennyiseg;
-            string partner;
-
-            public static List<Termek> keszlet = new(); //A termékeket reprezentáló struktúrák listája
-
-            /*Létrehozza a terméket reprezentáló struktúrát*/
-            public static Termek CreateProduct(List<KeszletMozgas> mozgas, string[] termekek, int index)
-            {
-                float[,] mennyisegek = KeszletMozgas.GetProductAmounts(mozgas);
-                string[,] beszallitok = KeszletMozgas.GetProductShippers(mozgas);
-
-                Termek termek = new()
-                {
-                    id = index + 1,
-                    ora = 16,
-                    perc = 00,
-                    tipus = MozgasTipus.BESZERZES,
-                    nev = termekek[index],
-                    mennyiseg = mennyisegek[index, (int)MozgasTipus.BESZERZES],
-                    partner = beszallitok[index, (int)MozgasTipus.BESZERZES]
-                };
-
-                return termek;
-            }
-
-            /*Összegzi a kifogyó termékeket és létrehozza a rendelési listát a megfelelő beszállítók hozzárendelésével*/
-            public static List<Termek> CreateOrderList(List<KeszletMozgas> mozgas)
-            {
-                string[] termekek = KeszletMozgas.OutOfStock(mozgas);
-                
-                int i = 0;
-                while (i < termekek.Length && termekek[i] is not null)
-                {
-                    keszlet.Add(CreateProduct(mozgas, termekek, i));
-                    i++;
-                }
-                return keszlet;
-            }
-
-            /*Létrehoz egy sort a rendelési listába a rendelési tétel adataiból*/
-            public static string CollectProductDetails(Termek termek)
-            {
-                string termekAdatok = $"{termek.id};{termek.ora}:{termek.perc};{termek.nev};{termek.mennyiseg};{termek.partner}";
-
-                return termekAdatok;
-            }
-
-        }
-
         /*Beolvassa és eltárolja a fájl tartalmát, ellenőrzi, hogy a fájlban vannak-e hibák, illetve üres sorok, amiről üzenetet is küld.*/
         static void LoadFile(string path, string output)
         {
